# Add and improve sqlite describe performance benchmarks