Add a few simple benchmarks for sqlite describe performance, then improve the benchmarks by replacing HashMap with a custom IntMap struct (a thin wrapper around Vec giving an API similar to HashMap<i32,_>).

```
insert/fks
----------
434836b4     1.00      21.2±0.05ms       ? ?/sec (IntMap)
0797aa13     1.30      27.6±0.09ms       ? ?/sec (BranchList)
0f18d2c0     1.54      32.5±0.12ms       ? ?/sec (separateMemory)
9c0fc8bf     1.54      32.6±0.10ms       ? ?/sec (initial)

insert/returning
----------------
434836b4     1.00     430.3±3.55µs       ? ?/sec (IntMap)
9c0fc8bf     1.08     464.6±4.59µs       ? ?/sec (initial)
0f18d2c0     1.09     468.1±4.92µs       ? ?/sec (separateMemory)
0797aa13     1.12     482.5±3.89µs       ? ?/sec (BranchList)

select/recursive
----------------
434836b4     1.00       7.8±0.02ms       ? ?/sec (IntMap)
9c0fc8bf     1.29      10.1±0.03ms       ? ?/sec (initial)
0f18d2c0     1.29      10.1±0.02ms       ? ?/sec (separateMemory)
0797aa13     1.50      11.8±0.03ms       ? ?/sec (BranchList)

select/trivial
--------------
434836b4     1.00     220.5±8.15µs       ? ?/sec (IntMap)
0f18d2c0     1.05     231.3±9.63µs       ? ?/sec (separateMemory)
9c0fc8bf     1.05     232.0±8.79µs       ? ?/sec (initial)
0797aa13     1.06     233.3±8.66µs       ? ?/sec (BranchList)
```